Fixed a problem with createNextTableOfContents starting with version 2
When a new ACS container was created and createNextTableOfContents was called,
the version returned would be 2 meaning an empty table of contents would exist
with a version of 1.  Now when an new ACS instance is created
ACS.getTableOfContents will return null and ACS.getCurrentVersion will return 0
until ACS.createNextTableOfContents is called the first time.