added sourceFile fallback on FileResourceIdentifier with a new constructor
When creating a FileResourceIdentifier the class will now first check
for an open sourceFileStream (FileInputStream) and use that, if that
doesn't exist then it will attempt to create a new one from the
sourceFile (File).  An IOException will be thrown if neither succeed.

A new constructor allows the creation of FileResourceIdentifiers
directly from a File, where before you could only use an
InputStream